Module Inf.Ns

A non-streamable implementation of the RFC 1950. It considers the input to be whole and is therefore able to save some time

type error = [

| `Invalid_header

| `Invalid_checksum

| De.Inf.Ns.error

]

The type for inflation errors.

val pp_error : Format.formatter -> error -> unit

Pretty-printer of error.

val inflate : bigstring -> bigstring -> (int * int, [> error ]) result

inflate src dst inflates the content of src into dst.

In case of success, it returns the bytes read and bytes writen in an Ok result. In case of failure, it returns the error in a Error result.